Conflicts:
	media/libvpx/vp8/encoder/arm/neon/vp8_shortwalsh4x4_neon.c

+++ browser/components/loop/content/js/panel.js
@@ -165,12 +165,11 @@ loop.panel = (function(_, mozL10n) {
   });
 
   var GettingStartedView = React.createClass({displayName: 'GettingStartedView',
-    componentDidMount: function() {
-      navigator.mozLoop.setLoopPref("gettingStarted.seen", true);
-    },
-
     handleButtonClick: function() {
-      navigator.mozLoop.openGettingStartedTour();
+      navigator.mozLoop.openGettingStartedTour("getting-started");
+      navigator.mozLoop.setLoopPref("gettingStarted.seen", true);
+      var event = new CustomEvent("GettingStartedSeen");
+      window.dispatchEvent(event);
     },
 
     render: function() {
@@ -215,7 +214,7 @@ loop.panel = (function(_, mozL10n) {
             )
           ),
         });
-        return React.DOM.div(null, 
+        return React.DOM.div({id: "powered-by-wrapper"}, 
           React.DOM.p({id: "powered-by"}, 
             mozL10n.get("powered_by_beforeLogo"), 
             React.DOM.img({id: "powered-by-logo", className: locale}), 
@@ -282,12 +281,19 @@ loop.panel = (function(_, mozL10n) {
       }
     },
 
+    handleHelpEntry: function(event) {
+      event.preventDefault();
+      var helloSupportUrl = navigator.mozLoop.getLoopPref('support_url');
+      window.open(helloSupportUrl);
+      window.close();
+    },
+
     _isSignedIn: function() {
       return !!navigator.mozLoop.userProfile;
     },
 
     openGettingStartedTour: function() {
-      navigator.mozLoop.openGettingStartedTour("settingsMenu");
+      navigator.mozLoop.openGettingStartedTour("settings-menu");
     },
 
     render: function() {
@@ -319,7 +325,10 @@ loop.panel = (function(_, mozL10n) {
                                           mozL10n.get("settings_menu_item_signin"), 
                                    onClick: this.handleClickAuthEntry, 
                                    displayed: navigator.mozLoop.fxAEnabled, 
-                                   icon: this._isSignedIn() ? "signout" : "signin"})
+                                   icon: this._isSignedIn() ? "signout" : "signin"}), 
+            SettingsDropdownEntry({label: mozL10n.get("help_label"), 
+                                   onClick: this.handleHelpEntry, 
+                                   icon: "help"})
           )
         )
       );
@@ -512,6 +521,65 @@ loop.panel = (function(_, mozL10n) {
     }
   });
 
+  var EditInPlace = React.createClass({displayName: 'EditInPlace',
+    mixins: [React.addons.LinkedStateMixin],
+
+    propTypes: {
+      onChange: React.PropTypes.func.isRequired,
+      text: React.PropTypes.string,
+    },
+
+    getDefaultProps: function() {
+      return {text: ""};
+    },
+
+    getInitialState: function() {
+      return {edit: false, text: this.props.text};
+    },
+
+    handleTextClick: function(event) {
+      event.stopPropagation();
+      event.preventDefault();
+      this.setState({edit: true}, function() {
+        this.getDOMNode().querySelector("input").select();
+      }.bind(this));
+    },
+
+    handleInputClick: function(event) {
+      event.stopPropagation();
+    },
+
+    handleFormSubmit: function(event) {
+      event.preventDefault();
+      this.props.onChange(this.state.text);
+      this.setState({edit: false});
+    },
+
+    cancelEdit: function(event) {
+      event.stopPropagation();
+      event.preventDefault();
+      this.setState({edit: false, text: this.props.text});
+    },
+
+    render: function() {
+      if (!this.state.edit) {
+        return (
+          React.DOM.span({className: "edit-in-place", onClick: this.handleTextClick, 
+                title: mozL10n.get("rooms_name_this_room_tooltip2")}, 
+            this.state.text
+          )
+        );
+      }
+      return (
+        React.DOM.form({onSubmit: this.handleFormSubmit}, 
+          React.DOM.input({type: "text", valueLink: this.linkState("text"), 
+                 onClick: this.handleInputClick, 
+                 onBlur: this.cancelEdit})
+        )
+      );
+    }
+  });
+
   /**
    * Room list entry.
    */
@@ -530,7 +598,7 @@ loop.panel = (function(_, mozL10n) {
         (nextState.urlCopied !== this.state.urlCopied);
     },
 
-    handleClickRoomUrl: function(event) {
+    handleClickEntry: function(event) {
       event.preventDefault();
       this.props.dispatcher.dispatch(new sharedActions.OpenRoom({
         roomToken: this.props.room.roomToken
@@ -538,6 +606,7 @@ loop.panel = (function(_, mozL10n) {
     },
 
     handleCopyButtonClick: function(event) {
+      event.stopPropagation();
       event.preventDefault();
       this.props.dispatcher.dispatch(new sharedActions.CopyRoomUrl({
         roomUrl: this.props.room.roomUrl
@@ -546,10 +615,31 @@ loop.panel = (function(_, mozL10n) {
     },
 
     handleDeleteButtonClick: function(event) {
+      event.stopPropagation();
       event.preventDefault();
-      // XXX We should prompt end user for confirmation; see bug 1092953.
-      this.props.dispatcher.dispatch(new sharedActions.DeleteRoom({
-        roomToken: this.props.room.roomToken
+      navigator.mozLoop.confirm({
+        message: mozL10n.get("rooms_list_deleteConfirmation_label"),
+        okButton: null,
+        cancelButton: null
+      }, function(err, result) {
+        if (err) {
+          throw err;
+        }
+
+        if (!result) {
+          return;
+        }
+
+        this.props.dispatcher.dispatch(new sharedActions.DeleteRoom({
+          roomToken: this.props.room.roomToken
+        }));
+      }.bind(this));
+    },
+
+    renameRoom: function(newRoomName) {
+      this.props.dispatcher.dispatch(new sharedActions.RenameRoom({
+        roomToken: this.props.room.roomToken,
+        newRoomName: newRoomName
       }));
     },
 
@@ -573,20 +663,19 @@ loop.panel = (function(_, mozL10n) {
       });
 
       return (
-        React.DOM.div({className: roomClasses, onMouseLeave: this.handleMouseLeave}, 
+        React.DOM.div({className: roomClasses, onMouseLeave: this.handleMouseLeave, 
+             onClick: this.handleClickEntry}, 
           React.DOM.h2(null, 
             React.DOM.span({className: "room-notification"}), 
-            room.roomName, 
+            EditInPlace({text: room.roomName, onChange: this.renameRoom}), 
             React.DOM.button({className: copyButtonClasses, 
+              title: mozL10n.get("rooms_list_copy_url_tooltip"), 
               onClick: this.handleCopyButtonClick}), 
             React.DOM.button({className: "delete-link", 
+              title: mozL10n.get("rooms_list_delete_tooltip"), 
               onClick: this.handleDeleteButtonClick})
           ), 
-          React.DOM.p(null, 
-            React.DOM.a({href: "#", onClick: this.handleClickRoomUrl}, 
-              room.roomUrl
-            )
-          )
+          React.DOM.p(null, React.DOM.a({href: "#"}, room.roomUrl))
         )
       );
     }
@@ -694,6 +783,7 @@ loop.panel = (function(_, mozL10n) {
     getInitialState: function() {
       return {
         userProfile: this.props.userProfile || navigator.mozLoop.userProfile,
+        gettingStartedSeen: navigator.mozLoop.getLoopPref("gettingStarted.seen"),
       };
     },
 
@@ -741,6 +831,23 @@ loop.panel = (function(_, mozL10n) {
       this.updateServiceErrors();
     },
 
+    _gettingStartedSeen: function() {
+      this.setState({
+        gettingStartedSeen: navigator.mozLoop.getLoopPref("gettingStarted.seen"),
+      });
+    },
+
+    _UIActionHandler: function(e) {
+      switch (e.detail.action) {
+        case "selectTab":
+          this.selectTab(e.detail.tab);
+          break;
+        default:
+          console.error("Invalid action", e.detail.action);
+          break;
+      }
+    },
+
     /**
      * The rooms feature is hidden by default for now. Once it gets mainstream,
      * this method can be simplified.
@@ -750,7 +857,6 @@ loop.panel = (function(_, mozL10n) {
         return (
           Tab({name: "call"}, 
             React.DOM.div({className: "content-area"}, 
-              GettingStartedView(null), 
               CallUrlResult({client: this.props.client, 
                              notifications: this.props.notifications, 
                              callUrl: this.props.callUrl}), 
@@ -762,7 +868,6 @@ loop.panel = (function(_, mozL10n) {
 
       return (
         Tab({name: "rooms"}, 
-          GettingStartedView(null), 
           RoomList({dispatcher: this.props.dispatcher, 
                     store: this.props.roomStore, 
                     userDisplayName: this._getUserDisplayName()}), 
@@ -786,10 +891,14 @@ loop.panel = (function(_, mozL10n) {
 
     componentDidMount: function() {
       window.addEventListener("LoopStatusChanged", this._onStatusChanged);
+      window.addEventListener("GettingStartedSeen", this._gettingStartedSeen);
+      window.addEventListener("UIAction", this._UIActionHandler);
     },
 
     componentWillUnmount: function() {
       window.removeEventListener("LoopStatusChanged", this._onStatusChanged);
+      window.removeEventListener("GettingStartedSeen", this._gettingStartedSeen);
+      window.removeEventListener("UIAction", this._UIActionHandler);
     },
 
     _getUserDisplayName: function() {
@@ -799,6 +908,17 @@ loop.panel = (function(_, mozL10n) {
 
     render: function() {
       var NotificationListView = sharedViews.NotificationListView;
+
+      if (!this.state.gettingStartedSeen) {
+        return (
+          React.DOM.div(null, 
+            NotificationListView({notifications: this.props.notifications, 
+                                  clearOnDocumentHidden: true}), 
+            GettingStartedView(null), 
+            ToSView(null)
+          )
+        );
+      }
 
       return (
         React.DOM.div(null,
